pyproject.toml
[project]
name = "hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "boto3>=1.37.13",
    "chainlit>=2.4.1",
    "pillow>=11.1.0",
    "python-dotenv>=1.0.1",
]
